refactor the trie verify cmd
