+++ verify/trie/types.go
@@ -1,0 +1,29 @@
+package trie
+
+import (
+	"github.com/NethermindEth/juno/core/crypto"
+	"github.com/NethermindEth/juno/core/trie"
+	"github.com/NethermindEth/juno/db"
+)
+
+const (
+	StarknetTrieHeight  uint8 = 251
+	ConcurrencyMaxDepth uint8 = 8
+)
+
+type TrieType string
+
+const (
+	ContractTrie        TrieType = "contract"
+	ClassTrie           TrieType = "class"
+	ContractStorageTrie TrieType = "contract-storage"
+)
+
+type TrieInfo struct {
+	Name       string
+	Prefix     []byte
+	HashFunc   trie.NewTrieFunc
+	HashFn     crypto.HashFn
+	ReaderFunc func(db.KeyValueReader, []byte, uint8) (trie.TrieReader, error)
+	Height     uint8
+}
